[sync] Move test to tests file and fix unwrap
