+++ sync/src/tasks/block_sync_task.rs
@@ -20,9 +20,9 @@ use stream_task::{CollectorState, TaskResultCollector, TaskState};
 
 #[derive(Clone, Debug)]
 pub struct SyncBlockData {
-    block: Block,
-    info: Option<BlockInfo>,
-    peer_id: Option<PeerId>,
+    pub(crate) block: Block,
+    pub(crate) info: Option<BlockInfo>,
+    pub(crate) peer_id: Option<PeerId>,
 }
 
 impl SyncBlockData {
@@ -287,199 +287,5 @@ where
 
     fn finish(self) -> Result<Self::Output> {
         Ok(self.chain)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use anyhow::format_err;
-    use futures::FutureExt;
-    use futures_timer::Delay;
-    use starcoin_accumulator::accumulator_info::AccumulatorInfo;
-    use starcoin_accumulator::tree_store::mock::MockAccumulatorStore;
-    use starcoin_accumulator::MerkleAccumulator;
-    use starcoin_crypto::HashValue;
-    use starcoin_types::block::BlockHeaderBuilder;
-    use starcoin_types::U256;
-    use std::collections::HashMap;
-    use std::sync::Mutex;
-    use std::time::Duration;
-    use stream_task::{Generator, TaskEventCounterHandle, TaskGenerator};
-
-    #[derive(Default)]
-    struct MockBlockFetcher {
-        blocks: Mutex<HashMap<HashValue, Block>>,
-    }
-
-    impl MockBlockFetcher {
-        pub fn new() -> Self {
-            Self::default()
-        }
-
-        pub fn put(&self, block: Block) {
-            self.blocks.lock().unwrap().insert(block.id(), block);
-        }
-    }
-
-    impl BlockFetcher for MockBlockFetcher {
-        fn fetch_block(
-            &self,
-            block_ids: Vec<HashValue>,
-        ) -> BoxFuture<Result<Vec<(Block, Option<PeerId>)>>> {
-            let blocks = self.blocks.lock().unwrap();
-            let result: Result<Vec<(Block, Option<PeerId>)>> = block_ids
-                .iter()
-                .map(|block_id| {
-                    if let Some(block) = blocks.get(block_id).cloned() {
-                        Ok((block, None))
-                    } else {
-                        Err(format_err!("Can not find block by id: {:?}", block_id))
-                    }
-                })
-                .collect();
-            async {
-                Delay::new(Duration::from_millis(100)).await;
-                result
-            }
-            .boxed()
-        }
-    }
-
-    fn build_block_fetcher(total_blocks: u64) -> (MockBlockFetcher, MerkleAccumulator) {
-        let fetcher = MockBlockFetcher::new();
-
-        let store = Arc::new(MockAccumulatorStore::new());
-        let accumulator = MerkleAccumulator::new_empty(store);
-        for i in 0..total_blocks {
-            let header = BlockHeaderBuilder::random().with_number(i).build();
-            let block = Block::new(header, vec![]);
-            accumulator.append(&[block.id()]).unwrap();
-            fetcher.put(block);
-        }
-        (fetcher, accumulator)
-    }
-
-    #[derive(Default)]
-    struct MockLocalBlockStore {
-        store: Mutex<HashMap<HashValue, SyncBlockData>>,
-    }
-
-    impl MockLocalBlockStore {
-        pub fn new() -> Self {
-            Self::default()
-        }
-
-        pub fn mock(&self, block: &Block) {
-            let block_id = block.id();
-            let block_info = BlockInfo::new(
-                block_id,
-                U256::from(1),
-                AccumulatorInfo::new(HashValue::random(), vec![], 0, 0),
-                AccumulatorInfo::new(HashValue::random(), vec![], 0, 0),
-            );
-            self.store.lock().unwrap().insert(
-                block.id(),
-                SyncBlockData::new(block.clone(), Some(block_info), None),
-            );
-        }
-    }
-
-    impl BlockLocalStore for MockLocalBlockStore {
-        fn get_block_with_info(
-            &self,
-            block_ids: Vec<HashValue>,
-        ) -> Result<Vec<Option<SyncBlockData>>> {
-            let store = self.store.lock().unwrap();
-            Ok(block_ids.iter().map(|id| store.get(id).cloned()).collect())
-        }
-    }
-
-    #[stest::test]
-    async fn test_block_sync() -> Result<()> {
-        let total_blocks = 100;
-        let (fetcher, accumulator) = build_block_fetcher(total_blocks);
-
-        let block_sync_state = BlockSyncTask::new(
-            accumulator,
-            0,
-            fetcher,
-            false,
-            MockLocalBlockStore::new(),
-            3,
-        );
-        let event_handle = Arc::new(TaskEventCounterHandle::new());
-        let sync_task =
-            TaskGenerator::new(block_sync_state, 5, 3, 1, vec![], event_handle.clone()).generate();
-        let result = sync_task.await?;
-        let last_block_number = result
-            .iter()
-            .map(|block_data| {
-                assert!(block_data.info.is_none());
-                block_data.block.header().number() as i64
-            })
-            .fold(-1, |parent, current| {
-                //ensure return block is ordered
-                assert_eq!(
-                    parent + 1,
-                    current,
-                    "block sync task not return ordered blocks"
-                );
-                current
-            });
-
-        assert_eq!(last_block_number as u64, total_blocks - 1);
-
-        let report = event_handle.get_reports().pop().unwrap();
-        debug!("report: {}", report);
-        Ok(())
-    }
-
-    #[stest::test]
-    async fn test_block_sync_with_local() -> Result<()> {
-        let total_blocks = 100;
-        let (fetcher, accumulator) = build_block_fetcher(total_blocks);
-
-        let local_store = MockLocalBlockStore::new();
-        fetcher
-            .blocks
-            .lock()
-            .unwrap()
-            .iter()
-            .for_each(|(_block_id, block)| {
-                if block.header().number() % 2 == 0 {
-                    local_store.mock(block)
-                }
-            });
-        let block_sync_state = BlockSyncTask::new(accumulator, 0, fetcher, true, local_store, 3);
-        let event_handle = Arc::new(TaskEventCounterHandle::new());
-        let sync_task =
-            TaskGenerator::new(block_sync_state, 5, 3, 1, vec![], event_handle.clone()).generate();
-        let result = sync_task.await?;
-        let last_block_number = result
-            .iter()
-            .map(|block_data| {
-                if block_data.block.header().number() % 2 == 0 {
-                    assert!(block_data.info.is_some())
-                } else {
-                    assert!(block_data.info.is_none())
-                }
-                block_data.block.header().number() as i64
-            })
-            .fold(-1, |parent, current| {
-                //ensure return block is ordered
-                assert_eq!(
-                    parent + 1,
-                    current,
-                    "block sync task not return ordered blocks"
-                );
-                current
-            });
-
-        assert_eq!(last_block_number as u64, total_blocks - 1);
-
-        let report = event_handle.get_reports().pop().unwrap();
-        debug!("report: {}", report);
-        Ok(())
     }
 }
