+++ sync/src/tasks/find_ancestor_task.rs
@@ -121,119 +121,3 @@ impl TaskResultCollector<BlockIdAndNumber> for AncestorCollector {
             .ok_or_else(|| format_err!("Unexpect state, collector finished by ancestor is None"))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::tasks::mock::MockBlockIdFetcher;
-    use logger::prelude::*;
-    use starcoin_accumulator::tree_store::mock::MockAccumulatorStore;
-    use starcoin_accumulator::{Accumulator, MerkleAccumulator};
-    use starcoin_crypto::HashValue;
-    use std::sync::Arc;
-    use stream_task::{Generator, TaskEventCounterHandle, TaskGenerator};
-
-    #[stest::test]
-    pub async fn test_find_ancestor_same_number() -> Result<()> {
-        let store = Arc::new(MockAccumulatorStore::new());
-        let accumulator = Arc::new(MerkleAccumulator::new_empty(store.clone()));
-
-        let fetcher = MockBlockIdFetcher::new(accumulator.clone());
-        fetcher.appends(generate_hash(100).as_slice())?;
-        let info0 = accumulator.get_info();
-
-        let store2 = Arc::new(MockAccumulatorStore::copy_from(store.as_ref()));
-        let accumulator2 = Arc::new(MerkleAccumulator::new_with_info(info0.clone(), store2));
-        let task_state = FindAncestorTask::new(
-            accumulator2.num_leaves() - 1,
-            accumulator.num_leaves() - 1,
-            7,
-            fetcher.clone(),
-        );
-        let event_handle = Arc::new(TaskEventCounterHandle::new());
-        let collector = AncestorCollector::new(accumulator2.clone());
-        let task =
-            TaskGenerator::new(task_state, 5, 3, 1, collector, event_handle.clone()).generate();
-        let ancestor = task.await?;
-        assert_eq!(ancestor.number, info0.num_leaves - 1);
-        let report = event_handle.get_reports().pop().unwrap();
-        debug!("report: {}", report);
-
-        Ok(())
-    }
-
-    #[stest::test]
-    pub async fn test_find_ancestor_block_number_behind() -> Result<()> {
-        let store = Arc::new(MockAccumulatorStore::new());
-        let accumulator = Arc::new(MerkleAccumulator::new_empty(store.clone()));
-
-        let fetcher = MockBlockIdFetcher::new(accumulator.clone());
-        fetcher.appends(generate_hash(100).as_slice())?;
-        let info0 = accumulator.get_info();
-
-        // remote node block id is greater than local.
-        fetcher.appends(generate_hash(100).as_slice())?;
-
-        let store2 = Arc::new(MockAccumulatorStore::copy_from(store.as_ref()));
-        let accumulator2 = Arc::new(MerkleAccumulator::new_with_info(info0.clone(), store2));
-        let task_state = FindAncestorTask::new(
-            accumulator2.num_leaves() - 1,
-            accumulator.num_leaves() - 1,
-            7,
-            fetcher.clone(),
-        );
-        let event_handle = Arc::new(TaskEventCounterHandle::new());
-        let collector = AncestorCollector::new(accumulator2.clone());
-        let task =
-            TaskGenerator::new(task_state, 5, 3, 1, collector, event_handle.clone()).generate();
-        let ancestor = task.await?;
-        assert_eq!(ancestor.number, info0.num_leaves - 1);
-        let report = event_handle.get_reports().pop().unwrap();
-        debug!("report: {}", report);
-
-        Ok(())
-    }
-
-    fn generate_hash(count: usize) -> Vec<HashValue> {
-        (0..count).map(|_| HashValue::random()).collect::<Vec<_>>()
-    }
-
-    #[stest::test]
-    pub async fn test_find_ancestor_chain_fork() -> Result<()> {
-        let store = Arc::new(MockAccumulatorStore::new());
-        let accumulator = Arc::new(MerkleAccumulator::new_empty(store.clone()));
-
-        let fetcher = MockBlockIdFetcher::new(accumulator.clone());
-        fetcher.appends(generate_hash(100).as_slice())?;
-        let info0 = accumulator.get_info();
-
-        fetcher.appends(generate_hash(100).as_slice())?;
-
-        let store2 = Arc::new(MockAccumulatorStore::copy_from(store.as_ref()));
-        let accumulator2 = Arc::new(MerkleAccumulator::new_with_info(info0.clone(), store2));
-
-        accumulator2.append(generate_hash(100).as_slice())?;
-        accumulator2.flush()?;
-
-        assert_ne!(accumulator.get_info(), accumulator2.get_info());
-        let batch_size = 7;
-        let task_state = FindAncestorTask::new(
-            accumulator2.num_leaves() - 1,
-            accumulator.num_leaves() - 1,
-            batch_size,
-            fetcher.clone(),
-        );
-        let event_handle = Arc::new(TaskEventCounterHandle::new());
-        let collector = AncestorCollector::new(accumulator2.clone());
-        let task =
-            TaskGenerator::new(task_state, 5, 3, 1, collector, event_handle.clone()).generate();
-        let ancestor = task.await?;
-        assert_eq!(ancestor.number, info0.num_leaves - 1);
-        let report = event_handle.get_reports().pop().unwrap();
-        debug!("report: {}", report);
-        //sub task not all finished, the collector return enough.
-        assert!(report.ok < report.sub_task);
-        assert_eq!(report.processed_items, 100 + 1);
-        Ok(())
-    }
-}
